Idkbro

+++ Assets/Scripts/GCScript.cs
@@ -13,8 +13,8 @@ public class GCScript : MonoBehaviour
     public Image healthMeter;
     public TMP_Text healthText2;
 
-    int health;
-    int maximumHealth = 100;
+    float health;
+    float maximumHealth = 100;
     float lerpSpeed;
 
     private void Start()
@@ -58,14 +58,14 @@ public class GCScript : MonoBehaviour
         healthMeter.color = healthMeterColor;
     }
 
-    public void Damage(int damageAmount)
+    public void Damage(float damageAmount)
     {
         if (health > 0)
         {
             health -= damageAmount;
         }
     }
-    public void Heal(int healAmount)
+    public void Heal(float healAmount)
     {
         if (health < maximumHealth)
         {
